pyproject.toml
[tool.poetry]
name = "whisper-server"
version = "0.1.0"
description = ""
authors = ["N <47500890+avi-cenna@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
faster-whisper = "^0.9.0"
loguru = "^0.7.2"
sounddevice = "^0.4.6"
webrtcvad = "^2.0.10"
pyzmq = "^25.1.1"
typer = "^0.9.0"
black = "^23.11.0"
isort = "^5.12.0"

[tool.black]
line-length = 100

[tool.isort]
profile = "black"
lines_between_types = 1

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
